// app-web/vela-wallet/src/lib/core/generated/DpermOperation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DpermGrant } from "./DpermGrant";
import type { DpermRespondPayload } from "./DpermRespondPayload";

export type DpermOperation = { "type": "write_grant", grant: DpermGrant, } | { "type": "remove_grant", origin: string, } | { "type": "respond", id: string, payload: DpermRespondPayload, } | { "type": "save_connection_record", address: string, chain_id: number, origin: string, };